Add helper script for mapping ports out of a FireFly stack
This script can be used to take a FireFly config file and replace
internal docker-compose ports with the mapped ports on localhost,
generating a new config file that can be exported.

It's most useful if you want to replace a container in a stack
with your own FireFly instance (often necessary during testing and
development on FireFly itself).

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>